src/mpc_planner/pyproject.toml
[tool.poetry]
name = "mpc_planner"
version = "0.1.0"
description = ""
authors = ["oscardegroot <groot.oscar.de@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.8.10"
casadi = "3.5.5"
numpy = ">=1.19.0, <1.20.0"
scipy = ">=1.4.1"
requests = ">=2.5"
pyyaml = "^6.0.1"
matplotlib = "3.6"
pytest-cov = "^4.1.0"
black = "^24.2.0"
acados-template = {path = "/workspace/acados/interfaces/acados_template", develop = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
